# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "KAI API",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login",
                "description": "get token after login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation with token and user's name",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register new user",
                "description": "register new user",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "username",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation message with user Token and user name",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "logout for both user and stylist",
                "description": "logout",
                "responses": {
                    "200": {
                        "description": "message : see you soon"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/forgot/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register new user",
                "description": "check email if exist and send reset email link",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "We have e-mailed your password reset link!",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "We can't find a user with that e-mail address."
                    }
                }
            }
        },
        "/api/find/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "find token password reset",
                "description": "find token password reset",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return json PasswordRest object ",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "This password reset token is invalid.."
                    }
                }
            }
        },
        "/api/email/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register new user",
                "description": "check email if exist and send reset email link",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "path",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user object",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "This password reset token is invalid."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickanme": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "introduction": {
                    "type": "string"
                },
                "photo_name": {
                    "type": "string"
                }
            }
        }
    }
}